Streamline naming convention and modularity of sphinx doc extension helpers for better readability. Introduce `example_path` as new parameter to `config_description`.
